moves `LodLiveUtils` and `shortenKey()` into a `utils` component
